+++ src/index.ts
@@ -1,11 +1,9 @@
-import fetch from "node-fetch";
+import { request } from "node:https";
+import { Train } from "./parsers/train";
+import { CTAResponse } from "./types/responses";
 
-import { buildUrl } from "./url";
-import {Train} from "./train";
-
-const BASE_URL = "http://lapi.transitchicago.com/api/1.0";
-const VERSION = require("../package.json").version;
-const USER_AGENT = `slow-zone/v${VERSION}`;
+const API_BASE_URL = "lapi.transitchicago.com";
+const API_BASE_PATH = "/api/1.0";
 
 export class SlowZone {
   apiKey: string;
@@ -15,58 +13,82 @@ export class SlowZone {
   }
 
   getArrivalsForStation(stationId, options = {}) {
-    return this.getArrivals({mapid: stationId, ...options});
+    return this.getArrivals({ mapid: stationId, ...options });
   }
 
   getArrivalsForStop(stopId, options = {}) {
-    return this.getArrivals({stpid: stopId, ...options});
+    return this.getArrivals({ stpid: stopId, ...options });
   }
 
   followTrain(runId) {
-    return this.fetch("ttfollow.aspx", {runnumber: runId});
-  }
-
-  // Private API
-  private fetch(endpoint, queryParams) {
-    const defaultQueryParams = {
-      key: this.apiKey,
-      outputType: "json",
-    };
-
-    return fetch(
-      buildUrl(BASE_URL, endpoint, {
-        ...queryParams,
-        ...defaultQueryParams,
-      }),
-      {
-        headers: {"User-Agent": USER_AGENT},
-      }
-    )
-      .then((res) => res.json())
-      .then(this.handleSuccess)
-      .catch(this.handleError);
+    return this.makeRequest("ttfollow.aspx", { runnumber: runId });
   }
 
   private getArrivals(options) {
-    return this.fetch("ttarrivals.aspx", options);
+    return this.makeRequest("ttarrivals.aspx", options);
   }
 
-  private handleSuccess(body) {
-    return new Promise((resolve, reject) => {
-      if (body.ctatt.errCd != "0") {
-        return reject(new Error(`${body.ctatt.errCd} – ${body.ctatt.errNm}`));
+  private async makeRequest(endpoint: string, queryParams = {}, data?) {
+    const query = new URLSearchParams({
+      ...queryParams,
+      key: this.apiKey,
+      outputType: "json",
+    });
+
+    const options = {
+      hostname: API_BASE_URL,
+      port: 443,
+      path: `${API_BASE_PATH}/${endpoint}?${query.toString()}`,
+      method: "GET",
+      headers: {
+        Accept: "application/json",
+        "User-Agent": "slow-zone/v3.0.0-beta1",
+      },
+    };
+
+    return new Promise(function (resolve, reject) {
+      const req = request(options, (res) => {
+        if (res.statusCode < 200 || res.statusCode >= 300) {
+          return reject(new Error(res.statusCode.toString()));
+        }
+
+        let body = [],
+          json;
+
+        res.on("data", (chunk) => {
+          body.push(chunk);
+        });
+
+        res.on("end", () => {
+          try {
+            json = JSON.parse(Buffer.concat(body).toString());
+          } catch (e) {
+            return reject(e);
+          }
+          return resolve(json);
+        });
+      });
+
+      req.on("error", (err) => {
+        reject(err);
+      });
+
+      if (data) {
+        req.write(data);
       }
 
-      const results = body.ctatt.eta.map((trainData) =>
-        new Train(trainData).toJSON()
-      );
+      req.end();
+    }).then((body: CTAResponse) => {
+      return new Promise((resolve, reject) => {
+        if (body.ctatt.errCd != "0") {
+          return reject(new Error(`[${body.ctatt.errCd}] ${body.ctatt.errNm}`));
+        }
 
-      resolve(results);
+        resolve(
+          body.ctatt.eta.map((trainData) => new Train(trainData).toJSON())
+        );
+      });
     });
-  }
-
-  private handleError(err) {
-    return Promise.reject(err);
   }
 }