+++ src/parsers/destination.ts
@@ -1,6 +1,9 @@
-const {getNativeInteger} = require("./helpers");
+import { getNativeInteger } from "./helpers";
+import type { TrainResponse } from "../types/responses";
 
-class Destination {
+export class Destination {
+  attributes: TrainResponse;
+
   constructor(attributes) {
     this.attributes = attributes;
   }
@@ -16,5 +19,3 @@ class Destination {
     };
   }
 }
-
-module.exports = Destination;