+++ test/index.js
@@ -1,0 +1,7 @@
+const SlowZone = require("../dist");
+
+describe("SlowZone", () => {
+  test("it exists", () => {
+    expect(SlowZone);
+  });
+});