+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.0.0
+
 - Migrate to TypeScript
 
 ## 2.1.6