+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 3.0.0
 
-- Migrate to TypeScript
+- Include types in package
+- Use HTTPS for communication with the API
+- Update documentation & links
+- Removed node-fetch dependency
 
 ## 2.1.6